# services/graph_harness_api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HarnessScheduleCreateRequest": {
        "description": "Create one continuous-learning schedule.",
        "properties": {
          "cadence": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Cadence",
            "type": "string"
          },
          "max_depth": {
            "default": 2,
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Max Depth",
            "type": "integer"
          },
          "max_new_proposals": {
            "default": 20,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Max New Proposals",
            "type": "integer"
          },
          "max_next_questions": {
            "default": 5,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Max Next Questions",
            "type": "integer"
          },
          "metadata": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JSONValue-Input"
            },
            "title": "Metadata",
            "type": "object"
          },
          "model_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "relation_types": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 200,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relation Types"
          },
          "run_budget": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HarnessRunBudget"
              },
              {
                "type": "null"
              }
            ]
          },
          "seed_entity_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 100,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed Entity Ids"
          },
          "source_type": {
            "default": "pubmed",
            "maxLength": 64,
            "minLength": 1,
            "title": "Source Type",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 256,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "cadence"
        ],
        "title": "HarnessScheduleCreateRequest",
        "type": "object"
      },
      "HarnessScheduleDetailResponse": {
        "description": "Schedule detail including recent runs.",
        "properties": {
          "recent_runs": {
            "items": {
              "$ref": "#/components/schemas/HarnessRunResponse"
            },
            "title": "Recent Runs",
            "type": "array"
          },
          "schedule": {
            "$ref": "#/components/schemas/HarnessScheduleResponse"
          }
        },
        "required": [
          "schedule",
          "recent_runs"
        ],
        "title": "HarnessScheduleDetailResponse",
        "type": "object"
      },
      "HarnessScheduleListResponse": {
        "description": "List response for schedule definitions.",
        "properties": {
          "schedules": {
            "items": {
              "$ref": "#/components/schemas/HarnessScheduleResponse"
            },
            "title": "Schedules",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "schedules",
          "total"
        ],
        "title": "HarnessScheduleListResponse",
        "type": "object"
      },
...
      "HarnessScheduleRunNowResponse": {
        "description": "Combined schedule state and immediate run result.",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/ContinuousLearningRunResponse"
          },
          "schedule": {
            "$ref": "#/components/schemas/HarnessScheduleResponse"
          }
        },
        "required": [
          "schedule",
          "result"
        ],
        "title": "HarnessScheduleRunNowResponse",
        "type": "object"
      },
...
      "JSONValue-Input": {
        "anyOf": [
          {
            "type": "string"
          },
          {
            "type": "integer"
          },
          {
            "type": "number"
          },
          {
            "type": "boolean"
          },
          {
            "additionalProperties": {
              "$ref": "#/components/schemas/JSONValue-Input"
            },
            "type": "object"
          },
          {
            "items": {
              "$ref": "#/components/schemas/JSONValue-Input"
            },
            "type": "array"
          },
          {
            "type": "null"
          }
        ]
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "MED13 Graph Harness API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/spaces/{space_id}/schedules": {
      "get": {
        "operationId": "list_schedules_v1_spaces__space_id__schedules_get",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HarnessScheduleListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List schedules",
        "tags": [
          "schedules"
        ]
      },
      "post": {
        "operationId": "create_schedule_v1_spaces__space_id__schedules_post",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HarnessScheduleCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HarnessScheduleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create schedule",
        "tags": [
          "schedules"
        ]
      }
    },
    "/v1/spaces/{space_id}/schedules/{schedule_id}": {
      "get": {
        "operationId": "get_schedule_v1_spaces__space_id__schedules__schedule_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HarnessScheduleDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get one schedule",
        "tags": [
          "schedules"
        ]
      },
      "patch": {
        "operationId": "update_schedule_v1_spaces__space_id__schedules__schedule_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HarnessScheduleUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HarnessScheduleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update schedule",
        "tags": [
          "schedules"
        ]
      }
    },
    "/v1/spaces/{space_id}/schedules/{schedule_id}/pause": {
      "post": {
        "operationId": "pause_schedule_v1_spaces__space_id__schedules__schedule_id__pause_post",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HarnessScheduleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Pause schedule",
        "tags": [
          "schedules"
        ]
      }
    },
    "/v1/spaces/{space_id}/schedules/{schedule_id}/resume": {
      "post": {
        "operationId": "resume_schedule_v1_spaces__space_id__schedules__schedule_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HarnessScheduleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Resume schedule",
        "tags": [
          "schedules"
        ]
      }
    },
    "/v1/spaces/{space_id}/schedules/{schedule_id}/run-now": {
      "post": {
        "operationId": "run_schedule_now_v1_spaces__space_id__schedules__schedule_id__run_now_post",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HarnessScheduleRunNowResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Trigger immediate schedule run",
        "tags": [
          "schedules"
        ]
      }
    }
  }
}